Move Vedeu::Model::Collection classes into one file.
