--- lib/vedeu/events/events.rb
@@ -1,10 +1,0 @@
-require 'vedeu/models/collection'
-
-module Vedeu
-
-  # A collection of {Vedeu::Event} instances.
-  #
-  class Events < Vedeu::Model::Collection
-  end # Events
-
-end # Vedeu
