+++ lib/vedeu/input/all.rb
@@ -1,6 +1,6 @@
 require 'vedeu/repositories/all'
 require 'vedeu/input/mapper'
-require 'vedeu/input/keys'
+require 'vedeu/models/view/collections'
 require 'vedeu/input/key'
 require 'vedeu/input/input'
 require 'vedeu/input/keymap'
