--- lib/vedeu/input/keys.rb
@@ -1,10 +1,0 @@
-require 'vedeu/models/collection'
-
-module Vedeu
-
-  # A collection of {Vedeu::Key} instances.
-  #
-  class Keys < Vedeu::Model::Collection
-  end # Keys
-
-end # Vedeu
