+++ lib/vedeu/models/view/all.rb
@@ -1,11 +1,10 @@
-require 'vedeu/models/view/chars'
+require 'vedeu/models/collection'
+require 'vedeu/models/view/collections'
+
 require 'vedeu/models/view/char'
 require 'vedeu/models/view/composition'
-require 'vedeu/models/view/interfaces'
 require 'vedeu/models/view/interface'
-require 'vedeu/models/view/lines'
 require 'vedeu/models/view/line'
-require 'vedeu/models/view/streams'
 require 'vedeu/models/view/stream'
 
 module Vedeu
