--- lib/vedeu/models/view/chars.rb
@@ -1,10 +1,0 @@
-require 'vedeu/models/collection'
-
-module Vedeu
-
-  # A collection of {Vedeu::Char} instances.
-  #
-  class Chars < Vedeu::Model::Collection
-  end # Chars
-
-end # Vedeu
