+++ lib/vedeu/models/view/collections.rb
@@ -1,0 +1,35 @@
+require 'vedeu/models/collection'
+
+module Vedeu
+
+  # A collection of {Vedeu::Char} instances.
+  #
+  class Chars < Vedeu::Model::Collection
+  end # Chars
+
+  # A collection of {Vedeu::Event} instances.
+  #
+  class Events < Vedeu::Model::Collection
+  end # Events
+
+  # A collection of {Vedeu::Interface} instances.
+  #
+  class Interfaces < Vedeu::Model::Collection
+  end # Interfaces
+
+  # A collection of {Vedeu::Key} instances.
+  #
+  class Keys < Vedeu::Model::Collection
+  end # Keys
+
+  # A collection of {Vedeu::Line} instances.
+  #
+  class Lines < Vedeu::Model::Collection
+  end # Lines
+
+  # A collection of {Vedeu::Stream} instances.
+  #
+  class Streams < Vedeu::Model::Collection
+  end # Streams
+
+end # Vedeu
