+++ lib/vedeu/models/view/interface.rb
@@ -4,7 +4,7 @@ require 'vedeu/output/presentation'
 require 'vedeu/buffers/display_buffer'
 require 'vedeu/buffers/buffer'
 
-require 'vedeu/models/view/lines'
+require 'vedeu/models/view/collections'
 require 'vedeu/models/view/line'
 
 module Vedeu
