--- lib/vedeu/models/view/interfaces.rb
@@ -1,10 +1,0 @@
-require 'vedeu/models/collection'
-
-module Vedeu
-
-  # A collection of {Vedeu::Interface} instances.
-  #
-  class Interfaces < Vedeu::Model::Collection
-  end # Interfaces
-
-end # Vedeu
