--- lib/vedeu/models/view/lines.rb
@@ -1,10 +1,0 @@
-require 'vedeu/models/collection'
-
-module Vedeu
-
-  # A collection of {Vedeu::Line} instances.
-  #
-  class Lines < Vedeu::Model::Collection
-  end # Lines
-
-end # Vedeu
