--- lib/vedeu/models/view/streams.rb
@@ -1,10 +1,0 @@
-require 'vedeu/models/collection'
-
-module Vedeu
-
-  # A collection of {Vedeu::Stream} instances.
-  #
-  class Streams < Vedeu::Model::Collection
-  end # Streams
-
-end # Vedeu
