+++ test/lib/vedeu/input/keymap_test.rb
@@ -46,7 +46,7 @@ module Vedeu
       subject { instance.key_defined?(input) }
 
       context 'when the input is defined' do
-        let(:keys) { Vedeu::Model::Collection.new([key]) }
+        let(:keys) { Vedeu::Keys.new([key]) }
 
         it { subject.must_be_instance_of(TrueClass) }
       end
@@ -65,7 +65,7 @@ module Vedeu
         let(:key_a) { Vedeu::Key.new('a') { :key_a } }
         let(:key_b) { Vedeu::Key.new('b') { :key_b } }
         let(:key_c) { Vedeu::Key.new('c') { :key_c } }
-        let(:keys)  { Vedeu::Model::Collection.new([key_a, key_b, key_c]) }
+        let(:keys)  { Vedeu::Keys.new([key_a, key_b, key_c]) }
 
         it { subject.must_be_instance_of(Array) }
         it { subject.must_equal([:key_b]) }
